QRandR removed, KScreen::config is working, setConfig yet to be ported
As KScreen turn out, there is no place for a huge XRandR wrapper as
QRandR, it makes our life way easier than using directly XRandR but it
adds an overhead in both, performance and complexity that is not
paying off.

So, instead we are going to use XRandR directly, at the moment just by
implementing big functions and later on splitting eveything up to have
a clean and easy to read code.

Next step will be making the setConfig code work, and then clean the
new code.