Added Generator, that will generate configs adjusted to the situation
Depending on different factors the ideal config is different, for example
depending if the device is a laptop, desktop, is docked, etc.

For that we need a class that will generate a config that works best with
every situation.

I don't like static classes, but it seemed to me like the most straight
forward design at this stage.