Store the mode info instead of the mode id, since we can't relay on it
The mode id may change as some test have demostrated alraedy, so we can't
relay on it to restore the config. Instead we are going to use the size
and the refresh and do a lookup ourselves to identify which mode we should
use.

The code is still ugly but hackeable, will make it better once it is
proven to work.