UPower docked property is not for what it seems it is, returning false
I assumed that UPower docked property was to know if the laptop was
sitting in a dock station but apparently it is only to know if there is
and external screen connected (like what we tried to do in the current
krandr kded module a few time ago).

For now I'm going to return always false and check if there is any other
way of doing it. If not I will just remove the feature for now.