Added Generator::displaySwitch(int iteration)
Laptop usually have a button used to switch between configurations for
example: extend, clone, etc...

In our case, this is the sequence of configurations we generate:

Clone
Extend to left
Turn off embedded (laptop)
Turn off external and turn on embedded
Extend to right

Extend to the right is the last option since we are defaulting to it when
a new output gets connected. Then we offer clone because of projectors.