Ship our own PlasmaComponents.Dialog
We switched from PlasmaCore.Dialog to PlasmaComponents.Dialog to
fix bug #312544. Unfortunatelly the component works correctly only
in KDE 4.10, it does not fix the bug in KDE 4.9 and it misbehaves.

To support KDE 4.9 (and older) and to make the code still working,
we ship our own copy of the Dialog.qml and it's dependencies from
kde-runtime/plasma/declarativeimports.

CCBUG: 312544