Use different cursors to indicate that outputs are draggable
Use Qt::OpenHandCursor when hovering anywhere on output
(except for buttons - fallback to Qt::ArrowCursor there)
and Qt::ClosedHandCursor when dragging the output around.

This is mostly a hack, because there is no way to change
cursor directly from QML, so we are through signals
changing cursor of the QDeclarativeView from C++ code.

BUG: 313958