do not rotate UI elements separate from screen
test subjects were unable to tell which side of the monitor was "up"
without prompting. the strip at the bottom of the screen was not clear.

when asked to rotate the screen into portrait, this was done without
much trouble. when asked to rotate it back ... they ended up putting
the screen upside down more often than not!

feedback noted that they thought it was right-side up because the
UI elements were right side up.