opacity is more obvious to denote "disabled"
not entirely happy with this, as the "on" switch also dims

however.. testers were unable to connect the shrinking of the monitor
with it being "off", even with the UI components being hidden. worse,
if there is another screen with the same resolution, then one has
two rectangles of the same size but one is lacking controls -> confusing.