Disable reaction when the lid is closed until we can do it properly
Modifying the Screen settings when the laptop is going to suspend is
quite annoying and stupid from our side since all we are going to
achieve is some flickering. Until we can know what will happen when
the lid closes we better disable this feature.

In the Solid sprint I hope we will be able to figure out an easy way
for kscreen to know when nothing will happen on lid close, and
reactivate this feature.