Fix crash when embeddedOutput() fails to detect embedded output
Some laptops (probably older models) use 'VGA-*' to identify embedded
outputs, so embeddedOutput() will fail, because it only looks for
modern 'LVDS', 'EDP', etc. In such case, just try to use the output
with lowest ID. It's highly probably that it will be the right one.

We should also try to come up with some really reliable method, instead
of name-based guessing.

BUG: 318907
FIXED-IN: 1.0