Port of the KScreen KCM to KF5
It is quite verbose as it was using some intrinsic API of QGraphicsView that
is obviously not available anymore.
It seems to work to some extent now, but it would be really good if there
was some testing since I'm unsure about how to test it.

CCMAIL: Dan Vrátil <dan.vratil@gmail.com>