Serializer: use qFuzzyCompare to match refresh rates
Refresh rates are floats. The rate we get from DBus (and thus from KScreen)
is correctly rounded to 6 decimal digits, but the number we get from
QVariant::toFloat() has arbitrary length, so operator == does not work here.
We could either round the numbers manually and compare them, but IMO it's
easier to just use qFuzzyCompare here.