QMLOutput: fix rounding error in updateRootProperties()
By storing the transformed coordinates as integers, we were loosing precision
for certain resolutions (like 1366x768), which caused the KCM to snap the outputs
inaccurately, and as a result, when calculating the final relative position of
the real outputs to apply, we would end up with a few pixels of overlap or gap
between them.

Through lucky choice we use scaling factor of 8, and most resolutions are
divisible by 8, so the bug did not manifest itself on most computers.