KCM: Don't create new UI Widget whenever KCMKScreen::configReady is called
When KCM is opened in System Settings, switching to another tab and back won't
delete and re-create the KCModule instance, but it will only call load(), which
in our case means re-fetching KScreen::Config. As a result, re-opening the
KScreen tab would instantiate a new Widget and put it into layout, breaking things.
Now we reuse an existing Widget (if any), and just assign the new KSCreen::Config.

BUG: 343829
FIXED-IN: 5.2.2