KDED: make mode lookup in Serializer::findOutput() more bullet-proof
Make sure that we can handle corrupted data in config fles (either because the
file itself is corrupted or because we have written nonsense in it) and don't
try to restore some messed up configuration.

The method will now also try harder to get a valid mode in case it fails to
match the data from config file with real KScreen::Config (for whatever reason)
and will fallback to preferred mode and biggest mode if necessary. This should
prevent us from leaving users with no screens in case of config file corruption.

CCBUG: 342675