Serializer: calculate screen size needed for the loaded config
When the configuration we apply from Serializer needs bigger XRandR Screen then
the currently set, we need to update the screen size as well. This can happen for
example when one of the screens in the loaded config is rotated by 90 degrees,
which means that it requires higher XRandR screen then currently available.
Without this the content of the output would be weirdly cropped.

BUG: 346239
FIXED-IN: 5.3.0