Handle two outputs having the same EDID
If one has two identical monitors, they end up with two entries in my
config with the same ID.

when calling findOutput it matches the first ID to the first output
then matches the second ID also to the first output

end result is both get positioned at the same place in mirror mode.

This patch ensures that in the event of a clash we still find the two
outputs independently making sure the same output isn't matched twice.

REVIEW:124700