fix interface name
I am not quite confident how this worked (best guess is target order in
cmake). upower_interface.h is not used anywhere, what is used is
freedesktop_interface.h (in device.cpp) an consequently the build can
fail on this file not being found.

Switch the generated header to the actually used name to avoid build
failures from not being able to find it.

e.g. http://tinyurl.com/pjjod85
the error happens and in another concurrent make job the interface is
built in the kded/ directory, so best guess is that this largely depends
on timing to go wrong.

REVIEW: 124918