Fix crash when exiting kscreen kcm in systemsettings
When using threaded rendering loop [1] systemsettings crashed when
clicking the back button. This crash is caused inside the shader effect
used by the OpacityMask.
QtQuick informed us also that something is wrong:
"QSGDefaultLayer::bind: ShaderEffectSource: 'recursive' must be set to true when rendering recursively."

The actual problem is that orientationPanelContainer and the OpacityMask
are children to the monitor item causing recursion. By making them
siblings to the orientationPanelContainer the problem is resolved. The
error message from QtQuick is no longer shown and systemsettings doesn't
crash any more.

BUG: 344651
FIXED-IN: 5.4.3
REVIEW: 125734

[1] QSG_RENDER_LOOP=threaded systemsettings5