make the serializer a bit more robust
Summary:
The daemon has some valid code pathes involving null configs. I've not
seen this particular part crashing, but it seems prudent to be a bit
more careful with the pointers we receive.

Also, debug message when writing fails, this seems to be a common
problem, and we should be able to debug this better when it happens.

Test Plan: autotests still pass

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1534