improve output identification
Summary:
When identical outputs (i.e. outputs with the same edid) are connected,
the serializer can't match the outputs solely by their id, it will most
likely return the wrong one and set that up.

In this case, we also compare the output's name with the name saved in
the config: this one is different for every monitor.

This change should allow to configure and restore setups with identical
screens connected.

BUG:325277

Test Plan: Only compile-tested so far. I'm running out of nightly hours.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2156