kscreen kcm: block changed signal after save
Summary:
When changing position (and potentially other properties) of the screen
setup, we receive delayed signals way after the sync operation finishes.
This enables the apply button after the config has been applied, leading
to the apply/discard confirmation dialog when closing (or navigating
away from) the kscreen kcm.

This patch disables propagation of the changed() signal at the highest
level, meaning that config widgets internally still get updated, but the
apply button only gets enabled after 1000 milliseconds. This magic value
is what I've figured out by testing. It's long enough to re-request edid
information on two outputs, but short enough to get in the way from
further changes.

I'm not super-happy with the timer, but I don't see any other way to fix
this glitch in the user experience. We don't really have a way on X11 to
know when all changes are flushed.

BUG:365467
FIXED-IN: 5.7.3

Test Plan: Tested systemsettings and kcmshell, with this patch, it works reliably and doesn't get in the way.

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2154