[kcm] Show output name in output preview
Summary:
When identical displays are connected, the identify overlay would display
the output name, while the preview in the kcm displays the display vendor
and model. There's no way to tell which display is which.

This patch adds a row in the output preview, it moves the vendor up,displays and displays the output name underneath (if not duplicated).

BUG:362570

Before: {F190822}

After: {F190823}

Test Plan:
started kcm with multiple displays connected, I can now distinguish them
based on the name of the connector

Reviewers: #plasma, graesslin

Reviewed By: #plasma, graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2248