show an OSD on display button press
Summary:
Show the plasma OSD with icon "preferences-desktop-display-randr" and
text when the display button has been pressed.

When only one displays is connected, "No External Display" is connected.
When more displays are connected, "Changing Screen Layout" is shown.

This is pretty bare-bones, it doesn't show what kind of setup it will
choose now, or next. This could be done by improving Generator, now,
however. The problem is that when we're applying a new config, displays
will flicker due to being mode-set, so the OSD isn't all that useful
since the screens will settle down.
It is useful in case of only one screen, since it gives feedback to a
button that otherwise just appears do nothing.

I think this could be enhanced in the future by giving the opportunity
to pick a layout. For 5.8, this brings a noticeable improvement and lays
the base for further work.

Test Plan: manually tested both cases (1 and 2 screens)

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: graesslin, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2718