add combobox for new output action
This adds the UI bits for the top combo box which allows to configure a
default action when a new, unknown display is connected.

It needs two things, writing of the config, and the implementation in
the kded module. The actual actions need review as well, should match
more closely the actual implementation, and check if everything desired
is "possible enough".