correct possibly invalid current mode
When we get config change events, the list of modes might have changed.
This happens for resizing virtual machine windows, for example). The
current mode may now be invalid, in that case we set the output to the
preferred mode (which in that case is the vm window's size).

cherry-picked from master 23006c1b9bc6657

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2155