apply config change after correcting invalid mode
The logic here to find and set the correct mode was OK, we just forgot
to actually apply the changes.

This change makes resizing plasmashell to the vm host window "kind of"
work. When the window gets bigger it works, when either axis gets
smaller, I'm getting a black screen (which can be corrected by making
the window just a little bit bigger and triggering another modeset).
This is a separate problem, however, since I can reproduce the exact same
behaviour just using xrandr.

This means, that we do now have all the bits and pieces in place for
handling changes of modes at runtime. I'm not closing this bug just yet,
since I want to give this some more testing exposure and possibly iron
out some wrinks.

CCBUG:356864