show osd later
This patch moves the OSD to a later point. As we're setting the mode,
the display changes, so we now wait a bit after setting, and then show
the OSD to indicate what happened. This uses some timer magic to delay
the OSD since we really don't want to start showing it while outputs
still change, ... but ... X11.