Fix KScreen KCM screen placement with scaled outputs
Summary:
The QML part of the KCM gets the output size manually as it wants to do
all translations itself.

We need to divide the mode by the scale as wayland wants everything in
normalised global co-ordinates.

i.e a screen to the right of a 2x output with a mode of 2000 pixels,
will have an X of 1000.

Test Plan:
Had a scaled output (set using kwin config)
Moved some about.
They appeared exactly where it should have done.

Reviewers: #plasma, sebas

Reviewed By: #plasma, sebas

Subscribers: sebas, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5636