Get output size with output-geometry() rather than mode size directly
Summary:
We can use output->geometry to get the current mode size with transform
applied simplifying a bunch of code, whilst being currently exactly the
same code.

This also applies the is->horizontal ? width : height to the
displaySwitch button. Fixing a bug if you had a rotated laptop
monitor then pressed the display switch button then it would be using the
wrong geometry.

----

Real reason for this patch is that if I make geometry() the definitive
call for getting output size with transformations I can then do all the
scaling support in a simple tidy one-line edit.

Reviewers: #plasma, sebas

Reviewed By: #plasma, sebas

Subscribers: sebas, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5659