Automatic scaling selection
Summary:
The generator code is run by the KDED when we don't have a profile for a
monitor setup.

If we don't have a config this sets the screen scaling to 2 when it has
an DPI > 1.5*normal DPI. The vertical resolution is also checked that it
matches what you'd find on a high resolution screen to hopefully avoid
issues with bogus EDID.

We may need to tweak this heuristic based on feedback, especially when
want to support fractional scaling.

Test Plan:
Removed configs.
Rebooted
Laptop screen looked like it should.
Normal screen was 1x

Reviewers: #plasma, broulik, sebas

Reviewed By: #plasma, broulik, sebas

Subscribers: sebas, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7482