refresh the config after it changed
When we get a config change signaled, the internal config may not have updated (when the wayland connection is in process). This means that we may miss changed data. Reloading the config when it changed and only then saving it makes it more robust.

I'm not 100% happy about this patch, but tried a lot of things to get the output updated when the backend changes, but it just doesn't happen. With this patch, at least the correct scale value gets written to the config file.

CCBUG:384733