track the config to monitor, save scale
Summary:
When doApplyConfig is called, we forgot to change the monitored config,
which means we weren't tracking changes anymore. This patch makes sure
that the new config becomes the monitored one in that case.

Also, save the scale property to the json config, we forgot to do that.

CCBUG:384733

Test Plan: Scale changes through kwin are now saved, they weren't before

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7912