Don't re-align outputs when QMLOutput in clone mode moves
Summary:
Changing resolution of the 'Unified output' makes it move around which
then triggers QMLScreen::qmlOutputMoved() which tries to snap and re-align
all the outputs. This leads to the KScreen Outputs being positioned next
to each other in the final KScreen Config that is uploaded to KDED,
instead of on top of each other, virtually breaking the 'Unify outputs'
feature.

Test Plan:
Click 'Unify outputs' and change resolution of the unified screen. Click apply.
Resolution changes on both screens and they are now are correctly repositioned
to be in clone mode.

Reviewers: sebas

Reviewed By: sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8390