hide primary display combo when not supported
Summary:
Some backends, for example the kwayland backend, do not support
setting the primary display. The backends already make their
feature set known so we can just hide it from the UI.

It's enough to hide it from the UI since setting it doesn't make
any difference, the libkscreen backend will simply ignore it
internally, so it's just honest to not show it.

BUG:390638

Test Plan:
ran kscreen kcm on both, X11 and wayland platforms, combo
and label are shown and hidden as expected

Reviewers: #plasma, dvratil, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: kpiwowarski, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10723