Load a fixed screen configuration if configured
With this patch, kscreen will prefer a file called fixed-config from
~/.local/share/kscreen if it exists. The idea is that a user (let's
call him Ben) can force a configuration to be loaded even if the
hardware is a bit wonky (or, complex).

Specifically, this makes the UTS data arena more workable even when
screens come and go accidentally, the configuration won't be changed
to something entirely made up but mostly be kept intact. With setups
with lots of screens this can easily happen when some scientist walks
over a cable while being in some kind of virtual reality. (It
happens to me literally all the time).

We've been testing this feature for over a year, I'm merging it now
since it makes the lives of one of our high-profile users a lot
easier, so as a maintainer I'm fine with bearing the costs of this
additional code-path, just in this case.

CCMAIL:ben.simons@uts.edu.au