Add applet with screen layouts and presentation mode
One of LiMux client's requirements is for display configuration to be easily accessible by mouse.
The OSD cannot be accessed by mouse, so this applet offers commonly used screen layouts in an
easily accessible place.

To keep the screen on during a presentation (when the application does not do that or the user is
actually demonstrating something on the machine itself) one needs to uncheck the non-userfriendly
labeled "Enable powermanagement" check box in the battery monitor. Since this is also affects "screen
setup", it is placed in this plasmoid as well.

The widget can be placed as an always-visible plasmoid in the panel or in System Tray where it would
only show if presentation mode is enabled or more then one screen is connected.

Differential Revision: https://phabricator.kde.org/D14855