Redraw the slider if modes have changed
Summary:
This can happen if a monitor was not configured at all, the KCM is shown
when it first plugged in and the user selects to not configure it.
In the UI it will appear with no Resolution/Refresh rate and wrong size in
the QML view.

Backport of D17658
Depends on D17685

Reviewers: #plasma, dvratil

Reviewed By: dvratil

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17686