Introduce secondary control channel of the KScreen daemon
Summary:
Kcm and daemon can only communicate a limited data set through the backend.
This feature patch adds an additional level of control by providing a one-way
control channel through output and configuration control files.

The first property to be controlled through this channel is the output
retention. Other information shared in the future might be if a certain
resolution or refresh rate was manually set by the user or is the result
of an algorithm, so that the daemon on next startup can decide to recompute
the optimal value.

Test Plan:
Autotests updated, such that they are controled for now to only use
indvidual retention.

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T10028

Differential Revision: https://phabricator.kde.org/D16992