Add a common directory for functionality being used by kcm and daemon
Summary:
Certain functionality will be shared between kcm and daemon in the future,
but is for now not extensive  enough to be put into libkscreen.

In order to still reduce code duplication add a common top-level directory.

Initially this directory contains the data root path logic and the Control
classed, such that the kcm can access it in the future as well.

Test Plan: Autotest passes.

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T10028

Differential Revision: https://phabricator.kde.org/D16993