[kded] Introduce Config wrapper class
Summary:
In order to consolidate the code and as a preparatory step for larger changes
introduce a wrapper class for KScreen::Config types. This class combines
information about the type and the previously static Serializer methods, such
that the Serializer class can be removed.

Test Plan: Autotest updated.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T10028

Differential Revision: https://phabricator.kde.org/D16989