[kcm] Provide graphical interface to replicate outputs
Summary:
This adds visual means to describe and manipulate the set of replicas of an
output.

Heuristics are included for proper repositioning of outputs. An alternative
approach would be to save former values, but for now this seems to work fine
in most cases.

At the moment the replication status is not yet saved via the KScreen daemon
to the configuration file what needs another presumably small patch to KScreen.

**Video:** https://www.youtube.com/watch?v=j__FJZK-A-0

{F7308512}
{F7308515}

Test Plan: Tested on X11 and Wayland with patches to libkscreen, KWayland and KWin.

Reviewers: #kwin, #plasma

Maniphest Tasks: T11222

Differential Revision: https://phabricator.kde.org/D23666