Polish Displays KCM UI
Summary:
This patch improves and polishes the new KScreen KCM's UI in a variety of ways. A picture
says a thousand words, so look at the pictures. :)

BUG: 348126
FIXED-IN: 5.17.0

Test Plan:
Due to an unrelated bug, my second screen isn't showing up in the visualization, so please excuse the fact that only one appears there in these screenshots:

{F7369458}

{F7369494}

All functionality that was working before is still working now.

Reviewers: romangg, #vdg, #kwin, GB_2

Reviewed By: romangg, #vdg, #kwin, GB_2

Subscribers: filipf, ndavis, GB_2, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23798