Revert "i18n fixes"
This reverts commit 6ae066c02573450e644d9abe532ceda8475d34cb.

The change breaks the 100 chars limit on source code set as part of the
Frameworks Coding Style.

If there is currently no other possibility to facilitate translations of long
text lines in QML without breaking this limit a temporary exception can be
made, but this needs to be discussed first and a plan must be drafted on how to
solve this conflict in the future.

To start such a discussion about short and long-term solution please create a
review on Phabricator for this change.

CCMAIL: pino@kde.org