Gracefully replace outputModel
This patch has it signal the disappearance of the outputModel and only then deletes it.
This gives the QML side of things a chance to clean up, specifically QQuickComboBox stores the model in a QVariant
and crashes if you just delete its backing model and then assign a new one.
Ideally the model instance itself would never be replaced but just reset and populated with new outputs.

Differential Revision: https://phabricator.kde.org/D24735