Fix laptop screen being off when opening the lid
Summary:
filePath() returns the full path to the config file. readFile() then prepends configsDirPath() to the path, leading to a garbled file path and making it impossible to load the right file. We should instead just pass id() + _lidOpen to readFile.

BUG: 371447

Test Plan:
Connected a monitor to my laptop. Closed and reopened the lid.
Before:
Laptop screen was off. Needed turning on in the KCM.
After:
Laptop screen is on.

Both X11 and Wayland

Reviewers: #plasma, romangg

Reviewed By: #plasma, romangg

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24719