[KCM] Revert change to reduce size of screens in visualization
Summary:
This was done in D23798, but in retrospect, that was a mistake. When more than one output
is connected, their representations in the visualization become tiny and their text gets
scrunched and elided. The old size worked better. So, let's revert to that.

BUG: 413142

Test Plan: Multi-screen display is much improved: {F7641754}

Reviewers: #vdg, #plasma, romangg

Reviewed By: #plasma, romangg

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24833