feat(kcm): add option for auto rotation
Summary:
This makes auto rotation configurable from the KScreen KCM. The auto rotation
feature is enabled when a rotation sensor is active and the output is internal,
otherwise the controls are disabled and we fall back to manual configuration.

An open question at this point is if there should be an additional option to
only rotate automatically when tablet mode is active or always. Currently it
always rotates automatically.

{F7820617}

Test Plan: Tested on laptops with and without rotation sensor.

Reviewers: #plasma

Subscribers: apol, aaronhoneycutt, sasch, ngraham, davidedmundson

Maniphest Tasks: T11475

Differential Revision: https://phabricator.kde.org/D26038