feat: control replication
Summary:
Adds API to read and write the replication source of an output to the control
file. This allows to remember specifically which output was the replication
source for another one since the windowing system might only get position
and size information about the replica-relation. This is the case on X11 and
will likely be handled in a similar way on Wayland.

Test Plan: Compiles.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26220