feat(kded): add orientation sensor
Summary:
With this patch KScreen controls rotation of internal outputs if an orientation
sensor is present and emitting rotation values.

The default behavior is auto-rotation but can be configured through control
files.

This functionality is developed for the Wayland session but might potentially
also work on X11, what needs to be tested more.

Test Plan: Screen of convertible with orientation sensor rotates.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel, davidedmundson

Tags: #plasma

Maniphest Tasks: T11475

Differential Revision: https://phabricator.kde.org/D26037