fix(kcm): use explicit Kirigami.Units over singleton from a theme
Summary:
units lowercase is defined by

PlasmaCore as a singleton (which we generally shouldn't be using)
Kirigami's desktop theme plugin, as a singleton. This plugin is only
loaded on KDE

Kirigami.Units is the correct import that is always defined.

BUG: 417316

Test Plan: opened KCM normally and with forced QT_QPA_PLATFORM

Reviewers: #plasma, apol, romangg

Reviewed By: #plasma, apol, romangg

Subscribers: davidre, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27442