fix(kded): watch config also when known config is applied
Summary:
When a known config is applied the doApplyConfig with the Config wrapper
override is called. In this case we still want to activate config watching
like when a config from the ConfigPtr override is applied.

Otherwise we are not be able to detect a control change until the first time
a new config is applied.

Test Plan:
Started with rotation in manual mode. Switched in KCM only from manual to
automatic mode without changing anything else. With the patch the daemon
notices this change, without it not.

Reviewers: #plasma, bshah

Reviewed By: #plasma, bshah

Subscribers: bshah, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28386